Found the Miner!
Deep within the darkest tunnels of the Marstone Silver Mine, the
almighty blue miner has been discovered by research teams! It is yet to
be determined if there are more miners within, or if this is the last
miner we will ever see in the Marstone Mine.

(added the miner mesh by Julius and its textures, but I would prefer
Dalton's. It's in the works, and should be done this term. Hopefully.)