Organizing Art and C++ Class Folders
Here is a list of changes made:
- The files are now organized to reduce clutter and confusion

Known Bugs:
- The .umap level will be ignored in the future commits, currently we
will be manually adding in the whitebox to playtest implementations, and
the level design made by Kyle will also be available to view our current
art assets

Programming Notes:
- For future reference, we will now be using UpperCamelCase when naming
our classes (this is common UE4 practice).
- Behavior tree tasks are divided in three categories: BTT(Tasks),
BTS(Services), BTD(Decorators) So we will be using this information to
use this type of naming convention. For Example I had changed the
BTService_FindTarget class to BTS_FindTarget.

!! Please ask me questions and give me feedback, and refer to the Trello
Board for the Task Tracker. I will walk you through what my plans will
be for that Friday for Frameworks class !!