First Pass on Ranged Enemy AI
Implemented a placeholder Projectile to spawn the static mesh for visual
prototyping.

Enemy can interchange the skeletal meshes, weapon placeholder mesh is
set up to attach to the Character Mesh Component.
Added a scenecomponent to set as the spawn point.
Added a lifetime to the bullets.

Will work on the projectile destroying onhit, and setting a number of
ammo the enemy has and how much bullets they can carry in their
magazine.

Enemy can shoot, movement of the project moves using the Projectile
Movement Component (utilizing the initial and max speed, and having a
velocity added to it)

Projectiles set not to bounce.