ART
Added New Level Pieces + Created Collisions for them.

Added Wood Clutter Pieces + Created Collisions for them.

Created a Cubemap HDR and applied it as the lighting for the level.

Created a placeholder Ground + Wall Textures for the level and applied
them to the level.

Created a Glowing Orb Blueprint to use with map navigation playtesting
later on.

Modified many uassets and the lantern blueprint.

Getting into doing a lighting pass soon.