main menu is currently completed
controls don't need to be set, the engine will determine between the two
what is being used. it just needs to be set up in input. I still don't
have luck with shadows, control rebinding is not worth the amount of
work it will take as it not an easy transition, audio sliders work as
intended currently, need to talk to darrel about adding music to test.